# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.CreateUserPayload:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  main.DeactivateUserResponse:
    properties:
      orphaned_projects:
        items:
          $ref: '#/definitions/store.ProjectRef'
        type: array
      user:
        $ref: '#/definitions/store.User'
    type: object
...
  main.ResetPasswordPayload:
    properties:
      password:
        type: string
    type: object
...
  main.UpdateUserPayload:
    properties:
      name:
        type: string
      role:
        type: string
    type: object
...
  store.ProjectRef:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  store.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      must_change_password:
        type: boolean
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for Trackforge application.
  title: Trackforge API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Users, oldest first. Filter with active=true|false; omit it for
        both. Anyone logged in can list active users (project admins need it to add
        people); only admins and the superadmin see deactivated ones, so for everyone
        else the filter is always active=true
      parameters:
      - description: Only active (true) or only deactivated (false) users
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.User'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create an admin or member. Admins may only create members (403
        otherwise). The new user must change the password on first login
      parameters:
      - description: New user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - users
  /users/{userID}:
    patch:
      consumes:
      - application/json
      description: Change name and/or role (admin or member). Admins may only rename
        members; changing a role is superadmin-only. The superadmin's role can't be
        changed here
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Fields to change
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /users/{userID}/deactivate:
    post:
      description: The user can no longer log in. Their history, memberships, project
        roles and assignments stay. You can't deactivate yourself, and admins may
        only deactivate members. orphaned_projects lists the projects where they were
        the only active project admin
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.DeactivateUserResponse'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Deactivate a user
      tags:
      - users
  /users/{userID}/make-superadmin:
    post:
      description: Superadmin only. The target, who must be active, becomes the superadmin
        and the caller becomes an admin
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Transfer superadmin
      tags:
      - users
  /users/{userID}/reactivate:
    post:
      description: Admins may only reactivate members
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Reactivate a user
      tags:
      - users
  /users/{userID}/reset-password:
    post:
      consumes:
      - application/json
      description: Set a temporary password. The user must change it on their next
        login. Admins may only reset members' passwords
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Temporary password
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ResetPasswordPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - BearerAuth: []
      summary: Reset a user's password
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: JWT from POST /auth/login. Paste it with the scheme included, e.g.
      "Bearer eyJhbGciOi..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"